Fix design & optimize code
