+++ app/src/main/res/drawable/ic_play_circle_filled_48dp.xml
@@ -1,5 +1,5 @@
-<vector android:height="24dp" android:tint="#979797"
+<vector android:height="48dp" android:tint="#979797"
     android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    android:width="48dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="#FF000000" android:pathData="M12,2C6.48,2 2,6.48 2,12s4.48,10 10,10 10,-4.48 10,-10S17.52,2 12,2zM10,16.5v-9l6,4.5 -6,4.5z"/>
 </vector>
