+++ app/src/main/res/layout/all_songs_fragment.xml
@@ -17,36 +17,45 @@
             android:layout_height="match_parent"
             >
 
+            <ss.com.bannerslider.Slider
+                android:id="@+id/slider"
+                android:layout_width="match_parent"
+                android:layout_height="@dimen/dp_200"
+                app:layout_constraintTop_toTopOf="parent"
+                app:slider_interval="10000"
+                app:slider_loopSlides="true"
+                />
+
             <TextView
                 android:id="@+id/textTitleMusic"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_alignParentStart="true"
                 android:layout_marginStart="@dimen/dp_8"
                 android:layout_marginTop="@dimen/dp_8"
                 android:text="@string/text_all_music"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/slider"
                 style="@style/Text16BoldWhite"
                 />
 
             <TextView
+                android:id="@+id/textView"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_alignParentEnd="true"
                 android:layout_centerVertical="true"
+                android:layout_marginBottom="@dimen/dp_8"
                 android:layout_marginEnd="@dimen/dp_8"
-                android:layout_marginTop="@dimen/dp_8"
                 android:text="@string/text_show_all"
+                app:layout_constraintBottom_toTopOf="@+id/recyclerMusic"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
                 style="@style/Text12AlphaWhite"
                 />
 
             <android.support.v7.widget.RecyclerView
                 android:id="@+id/recyclerMusic"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
+                android:layout_height="match_parent"
                 android:layout_marginTop="@dimen/dp_8"
                 app:layout_constraintTop_toBottomOf="@+id/textTitleMusic"
                 />
@@ -69,17 +78,19 @@
                 android:layout_height="wrap_content"
                 android:layout_alignParentEnd="true"
                 android:layout_centerVertical="true"
+                android:layout_marginBottom="@dimen/dp_8"
                 android:layout_marginEnd="@dimen/dp_8"
-                android:layout_marginTop="@dimen/dp_8"
                 android:text="@string/text_show_all"
+                app:layout_constraintBottom_toTopOf="@+id/recyclerAudio"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/recyclerMusic"
                 style="@style/Text12AlphaWhite"
                 />
+
             <android.support.v7.widget.RecyclerView
                 android:id="@+id/recyclerAudio"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
+                android:layout_height="match_parent"
                 android:layout_marginTop="@dimen/dp_8"
                 app:layout_constraintTop_toBottomOf="@+id/textTitleAudio"
                 />
