+++ app/src/main/res/layout/home_activity.xml
@@ -5,7 +5,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_gradient"
-    tools:context=".screen.home.HomeActivity">
+    tools:context=".screen.home.HomeActivity"
+    >
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
@@ -14,29 +15,28 @@
         android:background="?attr/colorPrimary"
         app:layout_constraintTop_toTopOf="parent"
         app:popupTheme="@style/AppTheme.PopupOverlay"
-        app:theme="@style/ToolbarColoredBackArrow" />
-
-    <ss.com.bannerslider.Slider
-        android:id="@+id/slider"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/dp_200"
-        app:layout_constraintTop_toBottomOf="@+id/toolbar"
-        app:slider_interval="10000"
-        app:slider_loopSlides="true" />
+        app:theme="@style/ToolbarColoredBackArrow"
+        />
 
     <android.support.design.widget.TabLayout
         android:id="@+id/tabLayout"
         android:layout_width="match_parent"
         android:layout_height="@dimen/dp_0"
-        android:layout_below="@id/slider"
-        app:layout_constraintTop_toBottomOf="@+id/slider" />
+        app:layout_constraintTop_toBottomOf="@+id/toolbar"
+        app:tabBackground="?attr/colorPrimary"
+        app:tabIndicatorColor="@android:color/white"
+        app:tabMode="scrollable"
+        app:tabSelectedTextColor="@android:color/white"
+        app:tabTextColor="@color/colorAccent"
+        />
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewPager"
         android:layout_width="match_parent"
         android:layout_height="@dimen/dp_0"
         app:layout_constraintBottom_toTopOf="@+id/menuBottom"
-        app:layout_constraintTop_toBottomOf="@+id/tabLayout" />
+        app:layout_constraintTop_toBottomOf="@+id/tabLayout"
+        />
 
     <android.support.design.widget.BottomNavigationView
         android:id="@+id/menuBottom"
@@ -47,6 +47,7 @@
         app:itemIconTint="@android:color/white"
         app:itemTextColor="@android:color/white"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:menu="@menu/bottom_menu_home" />
+        app:menu="@menu/bottom_menu_home"
+        />
 
 </android.support.constraint.ConstraintLayout>
