+++ app/src/main/res/layout/item_track_square.xml
@@ -28,7 +28,7 @@
             android:layout_marginBottom="@dimen/dp_8"
             android:layout_marginEnd="@dimen/dp_8"
             android:alpha="0.7"
-            android:src="@drawable/ic_play_circle_filled_24dp"
+            android:src="@drawable/ic_play_circle_filled_48dp"
             app:layout_constraintBottom_toBottomOf="@+id/imageSong"
             app:layout_constraintEnd_toEndOf="@+id/imageSong"
             />
